fix: wrong casting for uint16 -> uint256

+++ contracts/RiskEngine.sol
@@ -682,9 +682,8 @@ contract RiskEngine {
                 }
             }
         }
-
         // revert if the position does not require any collateral
-        if (required == 0) revert Errors.ZeroCollateralRequirement();        
+        if (required == 0) revert Errors.ZeroCollateralRequirement();
     }
 
     /// @notice Calculate the required amount of collateral for leg `index` for position `tokenId` accounting for its partner leg.
@@ -735,6 +734,7 @@ contract RiskEngine {
         int16 utilization
     ) internal view returns (uint256 required) {
         // if position is short, use sell collateral ratio
+
         if (isLong == 0) {
             // compute the sell collateral ratio, which depends on the pool utilization
             uint256 sellCollateral = _sellCollateralRatio(utilization);
@@ -955,7 +955,7 @@ contract RiskEngine {
     /// @param utilization The pool utilization of this collateral vault at the time the position is minted
     /// @return buyCollateralRatio The buy collateral ratio at `utilization`
     function _buyCollateralRatio(
-        uint16 utilization
+        uint256 utilization
     ) internal view returns (uint256 buyCollateralRatio) {
         // linear from BUY to BUY/2 between 50% and 90%
         // the buy ratio is on a straight line defined between two points (x0,y0) and (x1,y1):
